airflow/requirements.txt
apache-airflow-providers-apache-spark
pyspark==3.5.8
dbt-core==1.11.7
dbt-postgres==1.10.0
astronomer-cosmos